fix(coding-agent): bind image paste to Alt+V on WSL
Windows Terminal intercepts Ctrl+V as a text paste, so the keypress never
reached the app and image paste silently did nothing on WSL. Bind image
paste to Alt+V on WSL to match native Windows. Extract isWSL into a shared
utils/wsl.ts helper.
