+++ packages/coding-agent/src/utils/clipboard-image.ts
@@ -6,6 +6,7 @@ import { join } from "path";
 
 import { clipboard } from "./clipboard-native.ts";
 import { loadPhoton } from "./photon.ts";
+import { isWSL } from "./wsl.ts";
 
 export type ClipboardImage = {
 	bytes: Uint8Array;
@@ -138,19 +139,6 @@ function readClipboardImageViaWlPaste(): ClipboardImage | null {
 	}
 
 	return { bytes: data.stdout, mimeType: baseMimeType(selectedType) };
-}
-
-function isWSL(env: NodeJS.ProcessEnv = process.env): boolean {
-	if (env.WSL_DISTRO_NAME || env.WSLENV) {
-		return true;
-	}
-
-	try {
-		const release = readFileSync("/proc/version", "utf-8");
-		return /microsoft|wsl/i.test(release);
-	} catch {
-		return false;
-	}
 }
 
 /**
